Cart contoller done, product controller started

+++ src/EzGameMarket/WebApps/WebMVC/ViewModels/Pagination/PaginationInfo.cs
@@ -8,7 +8,7 @@ namespace WebMVC.ViewModels.Pagination
     public class PaginationInfo
     {
         public int ActualPage { get; set; }
-        public int TotalItemsCount { get; set; }
+        public long TotalItemsCount { get; set; }
         public int MaxPageNumber { get; set; }
         public int ItemsPerPage { get; set; }
         public string Previous { get; set; }
